LayoutTests/media/track/track-webvtt-tc015-positioning.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            

            var numberOfTrackTests = 3;
            
            function trackLoaded()
            {
                numberOfTracksLoaded++;
                if (numberOfTracksLoaded == numberOfTrackTests) {
                    testTrack(0);
                    testTrack(1);
                    testTrackError(2);
                }
            }
            
            function testTrack(i)
            {
                findMediaElement();
                var expected = 
                {
                    length : 4,
                    tests:
                    [
                        {
                            property : "textPosition",
                            values : [0, 50, 50, 100],
                        },
                    ],
                };
                testCues(i, expected);

                allTestsEnded();
            }

            function testTrackError(i)
            {
                findMediaElement();
                var expected = 
                {
                    length : 8,
                    tests:
                    [
                        {
                            property : "textPosition",
                            values : [50, 50, 50, 50, 50, 50, 50, 50],
                        },
                    ],
                };
                testCues(i, expected);

                allTestsEnded();
            }
        </script>
    </head>
    <body onload="enableAllTextTracks()">
        <p>Tests cue text position from settings.</p>
        <video>
            <track src="captions-webvtt/tc015-positioning.vtt" onload="trackLoaded()">
            <track src="captions-webvtt/tc015-positioning-ltr.vtt" onload="trackLoaded()">
            <track src="captions-webvtt/tc015-positioning-bad.vtt" onload="trackLoaded()">
        </video>
    </body>
</html>
