LayoutTests/media/track/track-webvtt-tc015-positioning-expected.txt

Tests cue text position from settings.


*** Testing text track 0
EXPECTED (cues.length == '4') OK
EXPECTED (cues[0].textPosition == '0') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[2].textPosition == '50') OK
EXPECTED (cues[3].textPosition == '100') OK

*** Testing text track 1
EXPECTED (cues.length == '4') OK
EXPECTED (cues[0].textPosition == '0') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[2].textPosition == '50') OK
EXPECTED (cues[3].textPosition == '100') OK

*** Testing text track 2
EXPECTED (cues.length == '8') OK
EXPECTED (cues[0].textPosition == '50') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[2].textPosition == '50') OK
EXPECTED (cues[3].textPosition == '50') OK
EXPECTED (cues[4].textPosition == '50') OK
EXPECTED (cues[5].textPosition == '50') OK
EXPECTED (cues[6].textPosition == '50') OK
EXPECTED (cues[7].textPosition == '50') OK
END OF TEST
